+++ sdk/boleto/boleto.js
@@ -40,6 +40,7 @@ class Boleto extends Resource {
      * @param line [string, default null]: generated Boleto line for payment. ex: '34191.09008 63571.277308 71444.640008 5 81960000000062'
      * @param barCode [string, default null]: generated Boleto bar-code for payment. ex: '34195819600000000621090063571277307144464000'
      * @param status [string, default null]: current Boleto status. ex: 'registered' or 'paid'
+     * @param transactionIds [list of strings]: ledger transaction ids linked to this Invoice (if there are more than one, all but the first are reversals or failed reversal chargebacks). ex: ['19827356981273']
      * @param created [string, default null]: creation datetime for the Boleto. ex: '2020-03-10 10:30:00.000'
      * @param ourNumber [string, default null]: Reference number registered at the settlement bank. ex:“10131474”
      *
@@ -49,7 +50,7 @@ class Boleto extends Resource {
                     due = null, fine = null, interest = null, overdueLimit = null, receiverName = null,
                     receiverTaxId = null, tags = null, descriptions = null, discounts = null, id = null,
                     fee = null, line = null, barCode = null, status = null,
-                    created = null, ourNumber = null
+                    transactionIds = null, created = null, ourNumber = null
                 }) {
         super(id);
         this.amount = amount;
@@ -74,6 +75,7 @@ class Boleto extends Resource {
         this.line = line;
         this.barCode = barCode;
         this.status = status;
+        this.transactionIds = transactionIds;
         this.created = created;
         this.ourNumber = ourNumber;
     }
@@ -134,7 +136,7 @@ exports.pdf = async function (id, { layout, hiddenFields, user } = {}) {
      *
      * Parameters (optional):
      * @param layout [string, default 'default']: Layout specification. Available options are 'default' and 'booklet'
-     * @param hiddenFields [list of strings, default null]: List of string fields to be hidden in Boleto pdf. ex: ["customerAddress"]
+     * @param hiddenFields [list of strings, default null]: List of string fields to be hidden in Boleto pdf. ex: ['customerAddress']
      * @param user [Organization/Project object]: Organization or Project object. Not necessary if starkbank.user was set before function call
      *
      * Return: