+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - 'payment' account type for Pix related resources
 - Transfer.description property to allow control over corresponding Transaction descriptions
 - 'link' property to Invoice resource
+- missing parameters to Boleto, DictKey, Invoice, Transaction, UtilityPayment and Workspace resources
 
 ## [2.6.0] - 2021-06-06
 ### Fixed